FWK-11
git-svn-id: http://svn.code.sf.net/p/jcaptcha/code/trunk@227 eaf2ec09-4c2e-0410-84d0-fe0bfca4863a

+++ jcaptcha/engine/src/java/com/octo/captcha/engine/image/gimpy/BasicListGimpyEngine.java
@@ -497,7 +497,7 @@ public class    BasicListGimpyEngine
                 new Integer(200), new Integer(100), "gimpybackgrounds");
         //BackgroundGenerator funkyBack = new FunkyBackgroundGenerator(new Integer(200), new Integer(100));
         FontGenerator shearedFont = new TwistedAndShearedRandomFontGenerator(
-                new Integer(30), new Integer(35));
+                new Integer(20), new Integer(25));
         //word2image 1
         com.octo.captcha.component.image.wordtoimage.WordToImage word2image =
                 new com.octo.captcha.component.image.wordtoimage.ComposedWordToImage(
